funktionale.State,... -> arrow.data.State
