+++ arrow-data/src/test/kotlin/arrow/data/StateTests.kt
@@ -1,5 +1,6 @@
 package arrow
 
+import arrow.syntax.collections.*
 import io.kotlintest.KTestJUnitRunner
 import io.kotlintest.matchers.shouldBe
 import org.junit.runner.RunWith
@@ -8,6 +9,8 @@ import org.junit.runner.RunWith
 class StateTests : UnitSpec() {
 
     private val addOne = State<Int, Int> { n -> n * 2 toT n }
+
+    val add1 = State { n: Int -> n + 1 toT n }
 
     init {
         "addOne.run(1) should return Pair(2, 1)" {
@@ -28,6 +31,32 @@ class StateTests : UnitSpec() {
 
         "addOne.runA(1) should return 1" {
             addOne.runA(1) shouldBe 1
+        }
+
+        "basic" {
+            add1.run(1) shouldBe (2 toT 1)
+        }
+
+        "traverse" {
+            val ns = (0..10).toList()
+            val x = ns.stateTraverse { add1 }
+            x.run(0).a shouldBe 11
+        }
+
+        "pure" {
+            val s1 = State().pure<String, Int>(1)
+            s1.run("foo") shouldBe ("foo" toT 1)
+        }
+
+        "get" {
+            val s1 = State().get<String>()
+            s1.run("foo") shouldBe ("foo" toT "foo")
+        }
+
+        "modify" {
+            val s1 = State().modify<String> { "bar" }
+            val s2 = State().set("bar")
+            s1.run("foo") shouldBe s2.run("foo")
         }
     }
 }
