+++ arrow-syntax/src/main/kotlin/arrow/syntax/collections/list.kt
@@ -3,6 +3,8 @@ package arrow.syntax.collections
 import arrow.syntax.function.toOption
 import arrow.Option
 import arrow.Some
+import arrow.State
+import arrow.map
 import arrow.data.Disjunction
 import arrow.data.map
 
@@ -39,3 +41,11 @@ fun <T, L, R> List<T>.disjuntionTraverse(f: (T) -> Disjunction<L, R>): Disjuncti
 }
 
 fun <L, R> List<Disjunction<L, R>>.disjunctionSequential(): Disjunction<L, List<R>> = disjuntionTraverse { it }
+
+fun <R, S, T> List<T>.stateTraverse(f: (T) -> State<S, R>): State<S, List<R>> = foldRight(State.pure(emptyList())) { i: T, accumulator: State<S, List<R>> ->
+    f(i).map(accumulator, ({ head: R, tail: List<R> ->
+        head prependTo tail
+    }))
+}
+
+fun <S, T> List<State<S, T>>.stateSequential(): State<S, List<T>> = stateTraverse { it }
