--- funktionale-state/build.gradle
@@ -1,17 +1,0 @@
-dependencies {
-    compile project(':arrow-data')
-    compile project(':arrow-syntax')
-
-    kapt project(':arrow-annotations-processor')
-    kaptTest project(':arrow-annotations-processor')
-    compileOnly project(':arrow-annotations-processor')
-    testCompileOnly project(':arrow-annotations-processor')
-
-    testCompile project(':arrow-test')
-
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlinVersion"
-    compile "org.jetbrains.kotlinx:kotlinx-coroutines-core:$kotlinxCoroutinesVersion"
-}
-
-apply from: rootProject.file('gradle/generated-kotlin-sources.gradle')
-apply plugin: 'kotlin-kapt'
