--- funktionale-state/gradle.properties
@@ -1,4 +1,0 @@
-# Maven publishing configuration
-POM_NAME=State Module
-POM_ARTIFACT_ID=funktionale-state
-POM_PACKAGING=jar
