--- funktionale-state/src/test/kotlin/org/funktionale/state/StateTests.kt
@@ -1,58 +1,0 @@
-/*
- * Copyright 2013 - 2017 Mario Arias
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * 	http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.funktionale.state
-
-import io.kotlintest.KTestJUnitRunner
-import io.kotlintest.matchers.shouldBe
-import arrow.UnitSpec
-import org.junit.runner.RunWith
-
-
-@RunWith(KTestJUnitRunner::class)
-class StateTest : UnitSpec() {
-
-    init {
-
-        val add1 = State { n: Int -> n + 1 to n }
-
-        "basic" {
-            add1.run(1) shouldBe (2 to 1)
-        }
-
-        "traverse" {
-            val ns = (0..10).toList()
-            val x = ns.stateTraverse { add1 }
-            x.run(0).first shouldBe 11
-        }
-
-        "pure" {
-            val s1 = State.pure<String, Int>(1)
-            s1.run("foo") shouldBe ("foo" to 1)
-        }
-
-        "get" {
-            val s1 = State.get<String>()
-            s1.run("foo") shouldBe ("foo" to "foo")
-        }
-
-        "modify" {
-            val s1 = State.modify<String> { "bar" }
-            val s2 = State.set("bar")
-            s1.run("foo") shouldBe s2.run("foo")
-        }
-    }
-}
